test: Products Controller 테스트 코드 완료 및 Service 작성 중
